+++ src/tailwind/form/CloseButton.tsx
@@ -20,7 +20,7 @@ export const CloseButton = forwardRef<HTMLButtonElement, CloseButtonProps>((
     onClick={onClick}
     className={clsx(
       'tw:opacity-50 tw:highlight:opacity-80 tw:transition-opacity',
-      'tw:rounded-md tw:focus-ring',
+      'tw:rounded-md tw:focus-ring tw:cursor-pointer',
       className,
     )}
     aria-label={label}